delphinus/requirements.txt
# requirements.txt
google-generativeai
python-dotenv
rich
deepdiff
